requirements.txt
 langchain_core
python-dotenv
streamlit
langchain_community
langchain
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
arxiv
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox
lancedb
PyMuPDF
sendgrid
Pillow